backend/requirements.txt
Flask
flask-pymongo
werkzeug
pyjwt
flask-cors
boto3
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
click==8.2.0
colorama==0.4.6
docxcompose==1.4.0
docxtpl==0.20.0
et_xmlfile==2.0.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2025.2
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
